Add plotting and bucket_tables to bucketers, closes #65
