+++ skorecard/bucketers/bucketers.py
@@ -5,13 +5,13 @@ from probatus.binning import AgglomerativeBucketer
 from probatus.utils import ApproximationWarning
 
 from typing import Union, List, Dict
-from .base_bucketer import BaseBucketer
+from skorecard.bucketers.base_bucketer import BaseBucketer
 from skorecard.bucket_mapping import BucketMapping, FeaturesBucketMapping
 from skorecard.utils import NotInstalledError, NotPreBucketedError
+from skorecard.reporting import build_bucket_table
 
 from sklearn.tree import DecisionTreeClassifier
 from sklearn.tree import _tree
-from sklearn.utils.validation import check_is_fitted
 
 try:
     from optbinning import OptimalBinning
@@ -111,6 +111,7 @@ class OptimalBucketer(BaseBucketer):
             y = y.values
 
         self.features_bucket_mapping_ = {}
+        self.bucket_tables_ = {}
         self.binners = {}
 
         for feature in self.variables:
@@ -243,6 +244,7 @@ class EqualWidthBucketer(BaseBucketer):
         self._verify_specials_variables(self.specials, X.columns)
 
         self.features_bucket_mapping_ = {}
+        self.bucket_tables_ = {}
 
         for feature in self.variables:
 
@@ -272,11 +274,12 @@ class EqualWidthBucketer(BaseBucketer):
                 missing_treatment=self.missing_treatment,
             )
 
-        return self
+            # Calculate the bucket table
+            self.bucket_tables_[feature] = build_bucket_table(
+                X, y, column=feature, bucket_mapping=self.features_bucket_mapping_.get(feature)
+            )
 
-    def transform(self, X):
-        """Transform X."""
-        return super().transform(X)
+        return self
 
 
 class AgglomerativeClusteringBucketer(BaseBucketer):
@@ -339,6 +342,7 @@ class AgglomerativeClusteringBucketer(BaseBucketer):
         self._verify_specials_variables(self.specials, X.columns)
 
         self.features_bucket_mapping_ = {}
+        self.bucket_tables_ = {}
 
         for feature in self.variables:
             ab = AgglomerativeBucketer(bin_count=self.n_bins)
@@ -368,11 +372,12 @@ class AgglomerativeClusteringBucketer(BaseBucketer):
                 missing_treatment=self.missing_treatment,
             )
 
-        return self
+            # Calculate the bucket table
+            self.bucket_tables_[feature] = build_bucket_table(
+                X, y, column=feature, bucket_mapping=self.features_bucket_mapping_.get(feature)
+            )
 
-    def transform(self, X):
-        """Transform X."""
-        return super().transform(X)
+        return self
 
 
 class EqualFrequencyBucketer(BaseBucketer):
@@ -438,6 +443,7 @@ class EqualFrequencyBucketer(BaseBucketer):
         self._verify_specials_variables(self.specials, X.columns)
 
         self.features_bucket_mapping_ = {}
+        self.bucket_tables_ = {}
 
         for feature in self.variables:
 
@@ -471,11 +477,12 @@ class EqualFrequencyBucketer(BaseBucketer):
                 missing_treatment=self.missing_treatment,
             )
 
-        return self
+            # Calculate the bucket table
+            self.bucket_tables_[feature] = build_bucket_table(
+                X, y, column=feature, bucket_mapping=self.features_bucket_mapping_.get(feature)
+            )
 
-    def transform(self, X):
-        """Transform X."""
-        return super().transform(X)
+        return self
 
 
 class DecisionTreeBucketer(BaseBucketer):
@@ -567,6 +574,7 @@ class DecisionTreeBucketer(BaseBucketer):
         self._verify_specials_variables(self.specials, X.columns)
 
         self.features_bucket_mapping_ = {}
+        self.bucket_tables_ = {}
         self.binners = {}
 
         for feature in self.variables:
@@ -625,11 +633,12 @@ class DecisionTreeBucketer(BaseBucketer):
                 missing_treatment=self.missing_treatment,
             )
 
-        return self
+            # Calculate the bucket table
+            self.bucket_tables_[feature] = build_bucket_table(
+                X, y, column=feature, bucket_mapping=self.features_bucket_mapping_.get(feature)
+            )
 
-    def transform(self, X):
-        """Transform X."""
-        return super().transform(X)
+        return self
 
 
 class OrdinalCategoricalBucketer(BaseBucketer):
@@ -735,6 +744,7 @@ class OrdinalCategoricalBucketer(BaseBucketer):
         self._verify_specials_variables(self.specials, X.columns)
 
         self.features_bucket_mapping_ = {}
+        self.bucket_tables_ = {}
 
         for var in self.variables:
 
@@ -791,11 +801,12 @@ class OrdinalCategoricalBucketer(BaseBucketer):
                 missing_treatment=self.missing_treatment,
             )
 
-        return self
+            # Calculate the bucket table
+            self.bucket_tables_[var] = build_bucket_table(
+                X, y, column=var, bucket_mapping=self.features_bucket_mapping_.get(var)
+            )
 
-    def transform(self, X):
-        """Transform X."""
-        return super().transform(X)
+        return self
 
 
 class UserInputBucketer(BaseBucketer):
@@ -853,37 +864,24 @@ class UserInputBucketer(BaseBucketer):
         if not isinstance(features_bucket_mapping, FeaturesBucketMapping):
             if not isinstance(features_bucket_mapping, dict):
                 raise TypeError("'features_bucket_mapping' must be a dict or FeaturesBucketMapping instance")
-            self.features_bucket_mapping = FeaturesBucketMapping(features_bucket_mapping)
+            self.features_bucket_mapping_ = FeaturesBucketMapping(features_bucket_mapping)
         else:
-            self.features_bucket_mapping = features_bucket_mapping
+            self.features_bucket_mapping_ = features_bucket_mapping
 
         # If user did not specify any variables,
         # use all the variables defined in the features_bucket_mapping
         self.variables = variables
         if variables == []:
-            self.variables = list(self.features_bucket_mapping.maps.keys())
-
-        self.is_fitted_ = True
+            self.variables = list(self.features_bucket_mapping_.maps.keys())
 
     def fit(self, X, y=None):
         """Init the class."""
-        return self
-
-    def transform(self, X, y=None):
-        """Transforms an array into the corresponding buckets fitted by the Transformer.
-
-        Args:
-            X (pd.DataFrame): dataframe which will be transformed into the corresponding buckets
-            y (array): target
-
-        Returns:
-            df (pd.DataFrame): df with transformed features
-        """
-        check_is_fitted(self)
-        X = self._is_dataframe(X)
-
+        # bucket tables can only be computed on fit().
+        # so a user will have to .fit() if she/he wants .plot_buckets() and .bucket_table()
+        self.bucket_tables_ = {}
         for feature in self.variables:
-            bucket_mapping = self.features_bucket_mapping.get(feature)
-            X[feature] = bucket_mapping.transform(X[feature])
-
-        return X
+            # Calculate the bucket table
+            self.bucket_tables_[feature] = build_bucket_table(
+                X, y, column=feature, bucket_mapping=self.features_bucket_mapping_.get(feature)
+            )
+        return self
